requirements.txt
requests==2.32.2
selenium==4.21.0
python-dotenv==1.0.1